stable version

+++ DAL/Models/MyDay.cs
@@ -7,10 +7,10 @@ namespace DAL.Models
         public Guid UserId { get; set; }
         public Guid MyDaysGroupId { get; set; }
         public DateOnly Date { get; set; }
-        public Int64 Year { get; set; }
-        public Int64 Month { get; set; }
-        public Int64 Day { get; set; }
-        public Int64 Number { get; set; }
-        public Int64 Volume { get; set; } = 100;
+        public int Year { get; set; }
+        public int Month { get; set; }
+        public int Day { get; set; }
+        public int Number { get; set; }
+        public int Volume { get; set; } = 100;
     }
 }
